Refactor TablesEditor to improve UI/UX
