--- src/components/ColumnEditor/components/index.ts
@@ -1,1 +1,0 @@
-export * from './EditableColumnEditor';
