--- src/components/ColumnEditor/index.ts
@@ -1,2 +1,0 @@
-export * from './ColumnEditor';
-export { editableColumnEditorsRegistry } from './components';
