+++ src/components/editors/FieldPicker/utils.ts
@@ -1,10 +1,7 @@
 import { Field } from '@grafana/data';
 import { getFieldTypeIcon } from '@grafana/ui';
 
-/**
- * Get Field Key
- */
-export const getFieldKey = (source: string | number, name: string): string => `${source}:${name}`;
+import { getFieldKey } from '@/utils';
 
 /**
  * Get Field Option
@@ -12,9 +9,9 @@ export const getFieldKey = (source: string | number, name: string): string => `$
  * @param source
  */
 export const getFieldOption = (field: Field, source: string | number) => ({
-  value: getFieldKey(source, field.name),
+  value: getFieldKey({ source, name: field.name }),
   fieldName: field.name,
-  label: getFieldKey(source, field.name),
+  label: getFieldKey({ source, name: field.name }),
   source,
   icon: getFieldTypeIcon(field),
 });
