+++ src/components/editors/TablesEditor/TablesEditor.styles.ts
