+++ src/components/editors/TablesEditor/components/TableEditor/TableEditor.styles.ts
@@ -3,6 +3,9 @@ import { GrafanaTheme2 } from '@grafana/data';
 
 export const getStyles = (theme: GrafanaTheme2) => {
   return {
+    sectionWrapper: css`
+      margin-bottom: ${theme.spacing(0.5)};
+    `,
     sectionContent: css`
       padding: ${theme.spacing(0.5, 1)};
     `,
