+++ src/components/editors/TablesEditor/components/TableEditor/components/ColumnsEditor/ColumnsEditor.tsx
@@ -1,17 +1,16 @@
-import { cx } from '@emotion/css';
 import { DataFrame } from '@grafana/data';
 import { Button, Icon, IconButton, InlineField, InlineFieldRow, Tag, useTheme2 } from '@grafana/ui';
 import { DragDropContext, Draggable, DraggingStyle, Droppable, DropResult, NotDraggingStyle } from '@hello-pangea/dnd';
 import { Collapse } from '@volkovlabs/components';
 import React, { useCallback, useMemo, useState } from 'react';
 
+import { CollapseTitle, FieldPicker } from '@/components';
 import { DEFAULT_COLUMN_APPEARANCE, DEFAULT_COLUMN_EDIT, TEST_IDS } from '@/constants';
 import { CellAggregation, CellType, ColumnConfig, ColumnFilterMode, ColumnPinDirection, FieldSource } from '@/types';
-import { reorder } from '@/utils';
+import { getFieldKey, reorder } from '@/utils';
 
-import { ColumnEditor } from '../ColumnEditor';
-import { FieldPicker } from '../FieldPicker';
 import { getStyles } from './ColumnsEditor.styles';
+import { ColumnEditor } from './components';
 
 /**
  * Get Item Style
@@ -153,7 +152,7 @@ export const ColumnsEditor: React.FC<Props> = ({ value: groups, name, onChange, 
           {(provided) => (
             <div {...provided.droppableProps} ref={provided.innerRef}>
               {items.map((item, index) => (
-                <Draggable key={item.field.name} draggableId={item.field.name} index={index}>
+                <Draggable key={getFieldKey(item.field)} draggableId={getFieldKey(item.field)} index={index}>
                   {(provided, snapshot) => (
                     <div
                       ref={provided.innerRef}
@@ -162,32 +161,28 @@ export const ColumnsEditor: React.FC<Props> = ({ value: groups, name, onChange, 
                       className={styles.item}
                     >
                       <Collapse
-                        headerTestId={TEST_IDS.columnsEditor.itemHeader.selector(item.field.name)}
-                        contentTestId={TEST_IDS.columnsEditor.itemContent.selector(item.field.name)}
+                        headerTestId={TEST_IDS.columnsEditor.itemHeader.selector(getFieldKey(item.field))}
+                        contentTestId={TEST_IDS.columnsEditor.itemContent.selector(getFieldKey(item.field))}
                         fill="solid"
                         title={
-                          <>
-                            {item.field.name && (
-                              <div className={styles.titleWrapper}>
-                                <div className={cx(styles.title)}>
-                                  {item.field.name}
-                                  {item.group && <Tag name="Group" />}
-                                  {item.edit.enabled && <Tag name="Editable" />}
-                                  {item.pin === ColumnPinDirection.LEFT && <Tag name="Pinned: Left" />}
-                                  {item.pin === ColumnPinDirection.RIGHT && <Tag name="Pinned: Right" />}
-                                  {item.filter.enabled && <Tag name="Filterable" />}
-                                  {item.sort.enabled && <Tag name="Sortable" />}
-                                </div>
-                              </div>
-                            )}
-                          </>
+                          <CollapseTitle>
+                            {item.field.name}
+                            {item.group && <Tag name="Group" />}
+                            {item.edit.enabled && <Tag name="Editable" />}
+                            {item.pin === ColumnPinDirection.LEFT && <Tag name="Pinned: Left" />}
+                            {item.pin === ColumnPinDirection.RIGHT && <Tag name="Pinned: Right" />}
+                            {item.filter.enabled && <Tag name="Filterable" />}
+                            {item.sort.enabled && <Tag name="Sortable" />}
+                          </CollapseTitle>
                         }
                         actions={
                           <>
                             <IconButton
                               name="trash-alt"
                               onClick={() =>
-                                onChangeItems(items.filter((column) => column.field.name !== item.field.name))
+                                onChangeItems(
+                                  items.filter((column) => getFieldKey(column.field) !== getFieldKey(item.field))
+                                )
                               }
                               aria-label="Remove"
                               {...TEST_IDS.columnsEditor.buttonRemove.apply()}
@@ -202,11 +197,11 @@ export const ColumnsEditor: React.FC<Props> = ({ value: groups, name, onChange, 
                             </div>
                           </>
                         }
-                        isOpen={expanded[item.field.name]}
+                        isOpen={expanded[getFieldKey(item.field)]}
                         onToggle={(isOpen) =>
                           setExpanded({
                             ...expanded,
-                            [item.field.name]: isOpen,
+                            [getFieldKey(item.field)]: isOpen,
                           })
                         }
                       >
@@ -214,7 +209,9 @@ export const ColumnsEditor: React.FC<Props> = ({ value: groups, name, onChange, 
                           value={item}
                           onChange={(item) => {
                             onChangeItems(
-                              items.map((column) => (column.field.name === item.field.name ? item : column))
+                              items.map((column) =>
+                                getFieldKey(column.field) === getFieldKey(item.field) ? item : column
+                              )
                             );
                           }}
                           data={data}
