+++ src/components/editors/TablesEditor/components/TableEditor/components/ColumnsEditor/components/ColumnEditor/ColumnEditor.test.tsx
