+++ src/components/editors/TablesEditor/components/TableEditor/components/ColumnsEditor/components/ColumnEditor/index.ts
@@ -1,0 +1,1 @@
+export * from './ColumnEditor';
