+++ src/components/editors/TablesEditor/components/TableEditor/components/ColumnsEditor/index.ts
