+++ src/components/editors/TablesEditor/components/TableEditor/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ColumnsEditor';
+export * from './TableUpdateEditor';
