+++ src/components/editors/TablesEditor/components/TableEditor/index.ts
