+++ src/components/editors/TablesEditor/components/index.ts
@@ -1,0 +1,2 @@
+export * from './TableEditor';
+export * from './TableEditor';
