+++ src/components/editors/TablesEditor/index.ts
