+++ src/components/ui/CollapseTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './CollapseTitle';
