+++ src/components/ui/FieldsGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './FieldsGroup';
