+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './CollapseTitle';
+export * from './FieldsGroup';
